rename videoin => video etc.
also rename videoinput.h and namespace etc.
